Rudimentarily got plots to work!
