+++ src/ResearchOS/plot_runner.py
@@ -1,7 +1,10 @@
+from typing import TYPE_CHECKING
+import os
 
+if TYPE_CHECKING:
+    from ResearchOS.PipelineObjects.plot import Plot
 
 from ResearchOS.code_runner import CodeRunner
-from ResearchOS.PipelineObjects.plot import Plot
 
 class PlotRunner(CodeRunner):
 
@@ -16,15 +19,14 @@ class PlotRunner(CodeRunner):
         if pl.is_matlab:
             if not self.matlab_loaded:
                 raise ValueError("MATLAB is not loaded.")            
-            fcn = getattr(self.eng, "PlotWrapper")
+            fcn = getattr(self.matlab_eng, "PlotWrapper")
         else:
             fcn = getattr(pl, pl.method)
 
         # 2. Convert the vr_values_in to the right format.
         if not is_batch:
             vr_vals_in = list(vr_values_in.values())
-            if self.num_inputs > len(vr_vals_in): # There's an extra input open.
-                vr_vals_in.append(info)
+            vr_vals_in.append(info) # Always include the info variable.
         else:
             # Convert the vr_values_in to the right format.
             vr_vals_in = []
@@ -36,9 +38,14 @@ class PlotRunner(CodeRunner):
         # - the input variables
         # - the plot function name
         # - the file path to save the plot to
-
+        dataset_parent_folder = os.path.dirname(self.dataset.dataset_path)
+        plots_folder = os.path.join(dataset_parent_folder, "plots")
+        save_path = os.path.join(plots_folder, pl.id + " " + pl.name)
+        if not os.path.exists(save_path):
+            os.makedirs(save_path)
+        save_path = os.path.join(save_path, self.node.name + f" ({self.node.id})")
         try:
-            fig_handle = fcn(pl.mfunc_name, vr_vals_in)
+            fig_handle = fcn(pl.mfunc_name, save_path, vr_vals_in)
         except PlotRunner.matlab.engine.MatlabExecutionError as e:
             if "ResearchOS:" not in e.args[0]:
                 print("'ResearchOS:' not found in error message, ending run.")
