+++ src/ResearchOS/variable.py
@@ -13,13 +13,14 @@ class Variable(ResearchObject):
     prefix: str = "VR"
     _initialized: bool = False
 
-    def __getitem__(self, key: str) -> Any:
-        """Get an attribute."""
-        return getattr(self, key)
+    def __getitem__(self, slice: tuple) -> Any:
+        """Store the slice of the Variable."""
+        self.slice = slice
+        return self
 
     def __init__(self, hard_coded_value: Any = all_default_attrs["hard_coded_value"], 
                 **kwargs):
         if self._initialized:
             return
         self.hard_coded_value = hard_coded_value
-        super().__init__(**kwargs)        
+        super().__init__(**kwargs)
